Moved method out of the component
